Skończenie interfejsu użytkownika

+++ script_PHP/serduszko.php
@@ -1,22 +1,30 @@
 <?php
+// Otwarcie sesji
 session_start();
 
+// Sprawdzenie czy użytkownik jest zalogowany
 if (isset($_SESSION['zalogowany'])) {
+    // Połączenie z bazą danych
     $polaczenie = mysqli_connect('localhost', 'kpqmmvzc_uzytkownik', 'Użytkownik123', 'kpqmmvzc_forum');
 
-    $odp = $_GET['odp'];
-    $uzytkownik = $_SESSION['id'];
+    // Sprawdzenie czy nie ma problemów z połączeniem
+    if (!mysqli_connect_errno()) {
+        // Zdefiniowanie zmiennych
+        $odp = $_GET['odp'];
+        $uzytkownik = $_SESSION['id'];
 
-    if ($_GET['dod'] == 1) {
-        $sql = "INSERT INTO serca (odpowiedz_id, uzytkownik_id) VALUES ($odp, $uzytkownik)";
-        mysqli_query($polaczenie, $sql);
-    } else if ($_GET['dod'] == 2) {
-        $sql = "DELETE FROM serca WHERE odpowiedz_id = $odp AND uzytkownik_id = $uzytkownik";
-        mysqli_query($polaczenie, $sql);
+        // Dodanie lub usunięcie polubienia
+        if ($_GET['dod'] == 1) {
+            // Dodanie polubienia
+            $sql = "INSERT INTO serca (odpowiedz_id, uzytkownik_id) VALUES ($odp, $uzytkownik)";
+            mysqli_query($polaczenie, $sql);
+        } else if ($_GET['dod'] == 2) {
+            // Usunięcie polubienia
+            $sql = "DELETE FROM serca WHERE odpowiedz_id = $odp AND uzytkownik_id = $uzytkownik";
+            mysqli_query($polaczenie, $sql);
+        }
     }
     
-} else {
-    header("Location: index.php");
 }
 
 ?>
